Use proper (older) version of BC api for 4.2.2

+++ src/api/java/buildcraft/api/power/ILaserTarget.java
@@ -28,7 +28,7 @@ public interface ILaserTarget {
 	 *
 	 * @param energy
 	 */
-	void receiveLaserEnergy(double energy);
+	void receiveLaserEnergy(float energy);
 
 	/**
 	 * Return true if the Tile Entity object is no longer a valid target. For
